Use Configurable<> instead of customize(). Moved AO2DConverter definition to the header file.
