+++ EventVisualisation/Workflow/src/AO2DConverter.cxx
@@ -9,7 +9,7 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-/// \file
+/// \file AO2DConverter.cxx
 /// \author Piotr Nowakowski
 
 #include "EveWorkflow/AO2DConverter.h"
@@ -17,9 +17,7 @@
 #include "DetectorsBase/GeometryManager.h"
 #include "DetectorsBase/Propagator.h"
 #include "DataFormatsParameters/GRPObject.h"
-#include "DataFormatsGlobalTracking/RecoContainer.h"
 #include "DataFormatsTPC/WorkflowHelper.h"
-#include "DataFormatsITSMFT/TopologyDictionary.h"
 #include "DetectorsCommonDataFormats/NameConf.h"
 #include "ITSBase/GeometryTGeo.h"
 #include "ITSMFTReconstruction/ClustererParam.h"
@@ -33,10 +31,6 @@
 #include "DataFormatsMCH/TrackMCH.h"
 #include "DataFormatsMCH/ROFRecord.h"
 #include "DataFormatsMCH/ClusterBlock.h"
-#include "MCHTracking/TrackParam.h"
-#include "Framework/AnalysisTask.h"
-#include <unistd.h>
-#include <climits>
 
 using namespace o2::event_visualisation;
 using namespace o2::framework;
@@ -45,98 +39,72 @@ using namespace o2::globaltracking;
 using namespace o2::tpc;
 using namespace o2::trd;
 
-void customize(std::vector<ConfigParamSpec>& workflowOptions)
-{
-  std::vector<o2::framework::ConfigParamSpec> options{
-    {"jsons-folder", VariantType::String, "jsons", {"name of the folder to store json files"}},
-  };
+#include "Framework/runDataProcessing.h"
 
-  std::swap(workflowOptions, options);
+void AO2DConverter::init(o2::framework::InitContext& ic)
+{
+  LOG(INFO) << "------------------------    AO2DConverter::init version " << mWorkflowVersion << "    ------------------------------------";
+  const auto grp = o2::parameters::GRPObject::loadFrom();
+  o2::base::GeometryManager::loadGeometry();
+  o2::base::Propagator::initFieldFromGRP();
+  mConfig.reset(new EveConfiguration);
+  mConfig->configGRP.solenoidBz = 5.00668f * grp->getL3Current() / 30000.;
+  mConfig->configGRP.continuousMaxTimeBin = grp->isDetContinuousReadOut(o2::detectors::DetID::TPC) ? -1 : 0; // Number of timebins in timeframe if continuous, 0 otherwise
+  mConfig->ReadConfigurableParam();
+
+  auto gm = o2::trd::Geometry::instance();
+  gm->createPadPlaneArray();
+  gm->createClusterMatrixArray();
+  mTrdGeo.reset(new o2::trd::GeometryFlat(*gm));
+  mConfig->configCalib.trdGeometry = mTrdGeo.get();
+
+  std::string dictFileITS = o2::itsmft::ClustererParam<o2::detectors::DetID::ITS>::Instance().dictFilePath;
+  dictFileITS = o2::base::NameConf::getAlpideClusterDictionaryFileName(o2::detectors::DetID::ITS, dictFileITS);
+  if (o2::utils::Str::pathExists(dictFileITS)) {
+    mITSDict.readFromFile(dictFileITS);
+    LOG(INFO) << "Running with provided ITS clusters dictionary: " << dictFileITS;
+  } else {
+    LOG(INFO) << "Dictionary " << dictFileITS << " is absent, ITS expects cluster patterns for all clusters";
+  }
+  mConfig->configCalib.itsPatternDict = &mITSDict;
+
+  std::string dictFileMFT = o2::itsmft::ClustererParam<o2::detectors::DetID::MFT>::Instance().dictFilePath;
+  dictFileMFT = o2::base::NameConf::getAlpideClusterDictionaryFileName(o2::detectors::DetID::MFT, dictFileMFT);
+  if (o2::utils::Str::pathExists(dictFileMFT)) {
+    mMFTDict.readFromFile(dictFileMFT);
+    LOG(INFO) << "Running with provided MFT clusters dictionary: " << dictFileMFT;
+  } else {
+    LOG(INFO) << "Dictionary " << dictFileMFT << " is absent, MFT expects cluster patterns for all clusters";
+  }
+  mConfig->configCalib.mftPatternDict = &mMFTDict;
+
+  o2::tof::Geo::Init();
+
+  o2::its::GeometryTGeo::Instance()->fillMatrixCache(
+    o2::math_utils::bit2Mask(o2::math_utils::TransformType::T2GRot,
+                             o2::math_utils::TransformType::T2G,
+                             o2::math_utils::TransformType::L2G,
+                             o2::math_utils::TransformType::T2L));
+
+  mCurrentEvent = 0;
 }
 
-#include "Framework/runDataProcessing.h" // main method must be included here (otherwise customize not used)
-
-std::string jsonPath;
-
-struct AO2DConverter
+void AO2DConverter::process(o2::aod::Tracks const& tracks)
 {
-  static constexpr float mWorkflowVersion = 1.00;
-
-  void init(o2::framework::InitContext& ic)
-  {
-    LOG(INFO) << "------------------------    AO2DConverter::init version " << mWorkflowVersion << "    ------------------------------------";
-    const auto grp = o2::parameters::GRPObject::loadFrom();
-    o2::base::GeometryManager::loadGeometry();
-    o2::base::Propagator::initFieldFromGRP();
-    mConfig.reset(new EveConfiguration);
-    mConfig->configGRP.solenoidBz = 5.00668f * grp->getL3Current() / 30000.;
-    mConfig->configGRP.continuousMaxTimeBin = grp->isDetContinuousReadOut(o2::detectors::DetID::TPC) ? -1 : 0; // Number of timebins in timeframe if continuous, 0 otherwise
-    mConfig->ReadConfigurableParam();
-
-    auto gm = o2::trd::Geometry::instance();
-    gm->createPadPlaneArray();
-    gm->createClusterMatrixArray();
-    mTrdGeo.reset(new o2::trd::GeometryFlat(*gm));
-    mConfig->configCalib.trdGeometry = mTrdGeo.get();
-
-    std::string dictFileITS = o2::itsmft::ClustererParam<o2::detectors::DetID::ITS>::Instance().dictFilePath;
-    dictFileITS = o2::base::NameConf::getAlpideClusterDictionaryFileName(o2::detectors::DetID::ITS, dictFileITS);
-    if (o2::utils::Str::pathExists(dictFileITS)) {
-      mITSDict.readFromFile(dictFileITS);
-      LOG(INFO) << "Running with provided ITS clusters dictionary: " << dictFileITS;
-    } else {
-      LOG(INFO) << "Dictionary " << dictFileITS << " is absent, ITS expects cluster patterns for all clusters";
-    }
-    mConfig->configCalib.itsPatternDict = &mITSDict;
-
-    std::string dictFileMFT = o2::itsmft::ClustererParam<o2::detectors::DetID::MFT>::Instance().dictFilePath;
-    dictFileMFT = o2::base::NameConf::getAlpideClusterDictionaryFileName(o2::detectors::DetID::MFT, dictFileMFT);
-    if (o2::utils::Str::pathExists(dictFileMFT)) {
-      mMFTDict.readFromFile(dictFileMFT);
-      LOG(INFO) << "Running with provided MFT clusters dictionary: " << dictFileMFT;
-    } else {
-      LOG(INFO) << "Dictionary " << dictFileMFT << " is absent, MFT expects cluster patterns for all clusters";
-    }
-    mConfig->configCalib.mftPatternDict = &mMFTDict;
-
-    o2::tof::Geo::Init();
-
-    o2::its::GeometryTGeo::Instance()->fillMatrixCache(
-      o2::math_utils::bit2Mask(o2::math_utils::TransformType::T2GRot,
-                               o2::math_utils::TransformType::T2G,
-                               o2::math_utils::TransformType::L2G,
-                               o2::math_utils::TransformType::T2L));
-
-    mCurrentEvent = 0;
+  EveWorkflowHelper helper;
+  for (auto& track : tracks) {
+    helper.drawAOD(track);
   }
 
-  void process(o2::aod::Tracks const& tracks)
-  {
-    EveWorkflowHelper helper;
-    for (auto& track : tracks) {
-      helper.drawAOD(track);
-    }
-
-    mCurrentEvent++;
-    helper.save(jsonPath, 1, {}, {}, mWorkflowVersion);
-  }
-
-  std::shared_ptr<EveWorkflowHelper> mHelper;
-  std::size_t mCurrentEvent;
-
-  o2::itsmft::TopologyDictionary mITSDict;
-  o2::itsmft::TopologyDictionary mMFTDict;
-  std::unique_ptr<EveConfiguration> mConfig;
-  std::unique_ptr<o2::trd::GeometryFlat> mTrdGeo;
-};
+  mCurrentEvent++;
+  helper.save(jsonPath, 1, {}, {}, mWorkflowVersion);
+}
 
 WorkflowSpec defineDataProcessing(ConfigContext const& cfgc)
 {
   LOG(INFO) << "------------------------    defineDataProcessing " << AO2DConverter::mWorkflowVersion << "    ------------------------------------";
 
-  jsonPath = cfgc.options().get<typeof(jsonPath)>("jsons-folder");
-
   return WorkflowSpec{
-    adaptAnalysisTask<AO2DConverter>(cfgc)
+    adaptAnalysisTask<AO2DConverter>(cfgc, TaskName{"o2-aodconverter"})
   };
 }
